убирает дребезг
