+++ index.html
@@ -236,6 +236,7 @@
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
 
+  <script src="./node_modules/lodash/lodash.js"></script>
   <script src="./nouislider/nouislider.js"></script>
   <script type="module" src="js/main.js"></script>
