+++ package.json
@@ -13,7 +13,8 @@
   "devDependencies": {
     "browser-sync": "2.26.14",
     "eslint": "^7.18.0",
-    "eslint-config-htmlacademy": "2.0.0"
+    "eslint-config-htmlacademy": "2.0.0",
+    "lodash": "^4.17.21"
   },
   "license": "MIT",
   "dependencies": {},
